opt: SEQUOIADBMAINSTREAM-4359 实现通用的 LightJob 机制
git-svn-id: http://192.168.20.11/sequoiadb/trunk@40407 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/include/ossPriorityQueue.hpp
@@ -56,10 +56,10 @@ private :
    mutable boost::mutex                            _mutex ;
    boost::condition_variable                       _cond ;
 public :
-   UINT32 size ()
+   UINT64 size ()
    {
       boost::mutex::scoped_lock lock ( _mutex ) ;
-      return (UINT32)_queue.size() ;
+      return (UINT64)_queue.size() ;
    }
    void push ( Data const& data )
    {
